+++ src/main/java/com/inq/wishhair/wesharewishhair/global/exception/ApiExceptionHandler.java
@@ -8,8 +8,8 @@ import com.inq.wishhair.wesharewishhair.review.controller.LikeReviewController;
 import com.inq.wishhair.wesharewishhair.review.controller.ReviewController;
 import com.inq.wishhair.wesharewishhair.review.controller.ReviewSearchController;
 import com.inq.wishhair.wesharewishhair.user.controller.*;
-import com.inq.wishhair.wesharewishhair.wishlist.controller.WishListController;
-import com.inq.wishhair.wesharewishhair.wishlist.controller.WishListSearchController;
+import com.inq.wishhair.wesharewishhair.hairstyle.controller.WishListController;
+import com.inq.wishhair.wesharewishhair.hairstyle.controller.WishListSearchController;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.HttpRequestMethodNotSupportedException;
 import org.springframework.web.bind.MethodArgumentNotValidException;
