+++ src/main/java/com/inq/wishhair/wesharewishhair/hairstyle/service/dto/response/WishListResponse.java
@@ -1,4 +1,4 @@
-package com.inq.wishhair.wesharewishhair.wishlist.service.dto.response;
+package com.inq.wishhair.wesharewishhair.hairstyle.service.dto.response;
 
 import com.inq.wishhair.wesharewishhair.hairstyle.domain.HairStyle;
 import com.inq.wishhair.wesharewishhair.hairstyle.service.dto.response.HashTagResponse;
