+++ src/test/java/com/inq/wishhair/wesharewishhair/global/base/ControllerTest.java
@@ -22,10 +22,10 @@ import com.inq.wishhair.wesharewishhair.review.service.ReviewService;
 import com.inq.wishhair.wesharewishhair.user.controller.*;
 import com.inq.wishhair.wesharewishhair.user.service.*;
 import com.inq.wishhair.wesharewishhair.hairstyle.controller.HairStyleController;
-import com.inq.wishhair.wesharewishhair.wishlist.controller.WishListController;
-import com.inq.wishhair.wesharewishhair.wishlist.controller.WishListSearchController;
-import com.inq.wishhair.wesharewishhair.wishlist.service.WishListSearchService;
-import com.inq.wishhair.wesharewishhair.wishlist.service.WishListService;
+import com.inq.wishhair.wesharewishhair.hairstyle.controller.WishListController;
+import com.inq.wishhair.wesharewishhair.hairstyle.controller.WishListSearchController;
+import com.inq.wishhair.wesharewishhair.hairstyle.service.WishListSearchService;
+import com.inq.wishhair.wesharewishhair.hairstyle.service.WishListService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
