+++ src/test/java/com/inq/wishhair/wesharewishhair/global/base/ServiceTest.java
@@ -7,8 +7,8 @@ import com.inq.wishhair.wesharewishhair.review.domain.ReviewRepository;
 import com.inq.wishhair.wesharewishhair.review.domain.likereview.LikeReviewRepository;
 import com.inq.wishhair.wesharewishhair.user.domain.UserRepository;
 import com.inq.wishhair.wesharewishhair.user.domain.point.PointRepository;
-import com.inq.wishhair.wesharewishhair.wishlist.domain.WishListRepository;
-import com.inq.wishhair.wesharewishhair.wishlist.domain.WishListSearchRepository;
+import com.inq.wishhair.wesharewishhair.hairstyle.domain.wishlist.WishListRepository;
+import com.inq.wishhair.wesharewishhair.hairstyle.domain.wishlist.WishListSearchRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.transaction.annotation.Transactional;
